requirements.txt
black==22.6.0
click==8.0.4
isort==5.10.1
mypy==0.971
pydocstyle==6.1.1
pylint==2.12.2
pytest==7.1.1
wheel==0.37.1
pyyaml==6.0
tabulate==0.8.9
numpy==1.22.3
scikit-learn==1.0.2
scikit-image==0.19.2
jupyter==1.0.0
pillow==9.0.1
termcolor==1.1.0
tensorboard==2.8.0
thop==0.0.31.post2005241907
pytorch-lightning==1.6.5
wandb==0.12.21

types-tabulate
types-PyYAML
types-protobuf
types-requests
types-setuptools
types-termcolor